change amount type

+++ suikotlin/src/main/java/io/cosmostation/suikotlin/SuiClient.kt
@@ -10,7 +10,7 @@ import io.cosmostation.suikotlin.key.SuiKey
 import io.cosmostation.suikotlin.model.*
 import io.cosmostation.suikotlin.secure.CipherHelper
 import org.bouncycastle.jce.provider.BouncyCastleProvider
-import java.math.BigInteger
+import java.math.BigDecimal
 import java.security.Security
 import java.util.*
 
@@ -94,7 +94,11 @@ class SuiClient {
         FaucetService.create().faucet(FixedAmountRequest(Recipient(address))).body()
 
     suspend fun transferSui(
-        objectId: String, receiver: String, sender: String, gasBudget: Int, amount: BigInteger? = null
+        objectId: String,
+        receiver: String,
+        sender: String,
+        gasBudget: Int,
+        amount: BigDecimal? = null
     ): SuiWrappedTxBytes? {
         val params = mutableListOf(sender, objectId, gasBudget, receiver)
         amount?.let { params.add(it) }
